No more “0.00000000347123 hours”

+++ secret_hitler.py
@@ -320,7 +320,7 @@ class Game(object):
 
     @staticmethod
     def format_time(seconds):
-        return "{0:0>2}h {0:0>2}m".format(seconds / 3600, (seconds / 60) % 60)
+        return "{0:0>2}h {0:0>2}m".format(int(seconds / 3600), int((seconds / 60) % 60))
 
     def show_time_logs(self):
         return "Time Logs:\n\n" + ("\n{}\n".format(self.show(["-"]))).join(
